+++ carbon_pools/mp_create_soil_C.py
@@ -80,7 +80,7 @@ def mp_create_soil_C(tile_id_list, run_date = None):
     print("Making mangrove soil C tiles...")
 
     # count/3 worked on a r4.16xlarge machine. Memory usage maxed out around 350 GB during the gdal_calc step.
-    pool = multiprocessing.Pool(processes=cn.count/3)
+    pool = multiprocessing.Pool(processes=int(cn.count/3))
     pool.map(create_soil_C.create_mangrove_soil_C, tile_id_list)
 
     # # For single processor use
@@ -98,7 +98,7 @@ def mp_create_soil_C(tile_id_list, run_date = None):
 
     print("Making mineral soil C tiles...")
 
-    pool = multiprocessing.Pool(processes=cn.count/2)
+    pool = multiprocessing.Pool(processes=int(cn.count/2))
     pool.map(create_soil_C.create_mineral_soil_C, tile_id_list)
 
     # # For single processor use
@@ -111,7 +111,7 @@ def mp_create_soil_C(tile_id_list, run_date = None):
     print("Making combined soil C tiles...")
 
     # With count/2 on an r4.16xlarge machine, this was overpowered (used about 240 GB). Could increase the pool.
-    pool = multiprocessing.Pool(processes=cn.count/2)
+    pool = multiprocessing.Pool(processes=int(cn.count/2))
     pool.map(create_soil_C.create_combined_soil_C, tile_id_list)
 
     # # For single processor use
